Move plugin settings factories into HasSettings trait
